fix: copy asset files (images) when imported
